+++ src/ConfigFactory.php
@@ -3,14 +3,26 @@
 /**
  * Title: iDEAL Advanced v3 config factory
  * Description:
- * Copyright: Copyright (c) 2005 - 2015
+ * Copyright: Copyright (c) 2005 - 2016
  * Company: Pronamic
+ *
  * @author Remco Tolsma
  * @version 1.0.0
  */
 class Pronamic_WP_Pay_Gateways_IDealAdvancedV3_ConfigFactory extends Pronamic_WP_Pay_GatewayConfigFactory {
+	private $config_class;
+
+	public function __construct( $config_class = 'Pronamic_WP_Pay_Gateways_IDealAdvancedV3_Config', $config_test_class = 'Pronamic_WP_Pay_Gateways_IDealAdvancedV3_Config' ) {
+		$this->config_class      = $config_class;
+		$this->config_test_class = $config_test_class;
+	}
+
 	public function get_config( $post_id ) {
-		$config = new Pronamic_WP_Pay_Gateways_IDealAdvancedV3_Config();
+		$mode = get_post_meta( $post_id, '_pronamic_gateway_mode', true );
+
+		$config_class = ( 'test' === $mode ) ? $this->config_test_class : $this->config_class;
+
+		$config = new $config_class();
 
 		$config->merchant_id = get_post_meta( $post_id, '_pronamic_gateway_ideal_merchant_id', true );
 		$config->sub_id      = get_post_meta( $post_id, '_pronamic_gateway_ideal_sub_id', true );
@@ -19,23 +31,6 @@ class Pronamic_WP_Pay_Gateways_IDealAdvancedV3_ConfigFactory extends Pronamic_WP
 		$config->private_key          = get_post_meta( $post_id, '_pronamic_gateway_ideal_private_key', true );
 		$config->private_key_password = get_post_meta( $post_id, '_pronamic_gateway_ideal_private_key_password', true );
 		$config->private_certificate  = get_post_meta( $post_id, '_pronamic_gateway_ideal_private_certificate', true );
-
-		$gateway_id = get_post_meta( $post_id, '_pronamic_gateway_id', true );
-		$mode       = get_post_meta( $post_id, '_pronamic_gateway_mode', true );
-
-		global $pronamic_pay_gateways;
-
-		$gateway  = $pronamic_pay_gateways[ $gateway_id ];
-		$settings = $gateway[ $mode ];
-
-		$url = $settings['payment_server_url'];
-
-		$config->url = $url;
-
-		$config->certificates = array();
-		foreach ( $gateway['certificates'] as $certificate ) {
-			$config->certificates[] = $certificate;
-		}
 
 		return $config;
 	}