+++ src/Country.php
@@ -3,8 +3,9 @@
 /**
  * Title: Country
  * Description:
- * Copyright: Copyright (c) 2005 - 2015
+ * Copyright: Copyright (c) 2005 - 2016
  * Company: Pronamic
+ *
  * @author Remco Tolsma
  * @version 1.0.0
  */
@@ -39,7 +40,7 @@ class Pronamic_WP_Pay_Gateways_IDealAdvancedV3_Country {
 	 *
 	 * @return string
 	 */
-	public function get_name( ) {
+	public function get_name() {
 		return $this->name;
 	}