+++ src/Transaction.php
@@ -3,8 +3,9 @@
 /**
  * Title: Transaction
  * Description:
- * Copyright: Copyright (c) 2005 - 2015
+ * Copyright: Copyright (c) 2005 - 2016
  * Company: Pronamic
+ *
  * @author Remco Tolsma
  * @version 1.0.0
  */
@@ -55,7 +56,7 @@ class Pronamic_WP_Pay_Gateways_IDealAdvancedV3_Transaction {
 	 *
 	 * @var string
 	 */
-	private $expirationPeriod;
+	private $expiration_period;
 
 	/**
 	 * Language
@@ -77,7 +78,7 @@ class Pronamic_WP_Pay_Gateways_IDealAdvancedV3_Transaction {
 	 *
 	 * @var string
 	 */
-	private $entranceCode;
+	private $entrance_code;
 
 	//////////////////////////////////////////////////